[chore] Switch to a modular folder structure
