+++ bright-events/src/Components/App/App.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.css';
+
+import Events from '../ViewEvents/Events';
+
+const App = () => {
+  return (
+    <div className="App">
+      <Events/>
+    </div>
+         );
+};
+
+export default App;
