+++ bright-events/src/Components/Login/Login.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import '../css/Register.css';
+import './Login.css';
 
 class Login extends Component {
   constructor(props){
