+++ bright-events/src/Components/Register/Register.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import '../css/Register.css';
+import './Register.css';
 
 class Register extends Component {
   constructor(props){
