+++ bright-events/src/Components/ViewEvents/styles.css
@@ -1,0 +1,33 @@
+html,
+body{
+    font-family: 'Lato', sans-serif;
+    background-image:url('https://image.ibb.co/hF9ENw/climate_kic_350836.jpg');
+    background-position: center;
+    background-color: #43304C;
+    background-repeat: no-repeat;
+    background-size: cover;
+
+}
+.container{
+    padding: 0 70px;
+}
+.navbar{
+    margin-bottom: 0;
+    background-color: #6B8175;
+    border: 0;
+    border-radius: 0;
+    font-size: 18px !important;
+    letter-spacing: 1px;
+}
+.navbar li a,.navbar .navbar-brand{
+    color: #fff !important;
+    background-color: #43304C !important;
+        
+}
+.navbar-default .navbar-toggle{
+    background-color: transparent;
+    color: #fff;
+}
+.jumbotron{
+    opacity: 0.8;
+}
