--- bright-events/src/static/js/App.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import '../css/App.css';
-import HelloWorldList from './HelloWorldList';
-import Register from './Register';
-import AboutUs from './AboutUs';
-import Login from './Login';
-import Events from './Events';
-
-const App = () => {
-  return (
-    <div className="App">
-      <Events/>
-    </div>
-         );
-};
-
-export default App;
